QR_API_0.0.1/requirements.txt
ez_setup
fastapi==0.75.0
pymupdf
phpserialize==1.3
python-multipart
opencv-python
opencv-contrib-python
cmake
pycryptodomex
pyzbar==0.1.9
uvicorn==0.17.6
gunicorn==20.1.0